Refactoring

+++ src/main/java/de/revout/pi/vplotter/main/DriverMoveObserverIf.java
@@ -4,7 +4,7 @@ import de.revout.pi.vplotter.converter.Pair;
 
 public interface DriverMoveObserverIf {
 
-	public void currentMove(int paramState, Pair paramToPoint, int paramStepCount, int paramActualStep);
+	public void currentMove(int paramState, Pair paramToPoint, long paramStepCount, long paramActualStep);
 	public void finish();
 	public void init();
